pyproject.toml
[project]
name = "health"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pytest>=8.4.2",
    "python-dotenv>=1.1.1",
    "should>=0.5.4",
]
